+++ src/Views/articles/show.html.php
@@ -1,3 +1,6 @@
+<?php
+session_start();
+?>
 <header>
     <div class="container">
         <div class="row">
@@ -28,15 +31,18 @@
         <blockquote>
             <em><?= $commentaire['content'] ?></em>
         </blockquote>
+    <?php if (isset($_SESSION['isAdmin']) && $_SESSION['isAdmin'] == 1) :?>
         <a href="index.php?controller=comment&task=delete&id=<?= $commentaire['id'] ?>" onclick="return window.confirm
         (`Êtes vous
         sûr de
         vouloir supprimer ce commentaire ?!`)">Supprimer</a>
+    <?php endif ?>
     <?php endforeach ?>
 <?php endif ?>
+<h3>Vous voulez réagir ? N'hésitez pas les bros !</h3>
 
+<?php if (isset($_SESSION['isAdmin'])) : ?>
 <form action="index.php?controller=comment&task=insert" method="POST">
-    <h3>Vous voulez réagir ? N'hésitez pas les bros !</h3>
     <label>Pseudo :</label>
     <input type="text" name="author" placeholder="Votre pseudo !">
     <label>Commentaire :</label>
@@ -44,3 +50,6 @@
     <input type="hidden" name="article_id" value="<?= $article_id ?>">
     <button>Commenter !</button>
 </form>
+<?php else : ?>
+<p><a href="index.php?controller=user&task=login" >Connectez-vous pour réagir<a/></p>
+<?php endif ?>
